requirements.txt
torch
torchvision
opencv-python
numpy
pytesseract
Pillow
transformers
easyocr
TTS
ipython
streamlit